requirements.txt
six
wheel
cachetools
cryptography
requests
asgiref
pytz
sqlparse
Django
django-postgresql-manager
django-swingtime
django-scheduler
django-crispy-forms
django-webpack-loader
django-cors-headers
django-markdown-deux
django-pagedown
django-material
django-timezone-field
djangorestframework
djangorestframework-simplejwt
django-user-accounts
django-webtest
django-debug-toolbar
django-allauth
django-oauth-toolkit
PyJWT
markdown2
ipregistry
Pillow
pinax-templates
lxml
pyyaml
uritemplate
sphinx
psycopg2
